Data type

+++ 14-binary_tree_balance.c
@@ -7,10 +7,10 @@
  *
  * Return: Balance factor of the tree, if tree is NULL return 0
  */
-size_t binary_tree_balance(const binary_tree_t *tree)
+int binary_tree_balance(const binary_tree_t *tree)
 {
-	size_t height_left = 0, height_right = 0;
-	size_t balance = 0;
+	int height_left = 0, height_right = 0;
+	int balance = 0;
 
 	if (tree == NULL)
 		return (0);
